refactor:move schemas to separate module
